+++ packages/deploy/src/contracts.rs
@@ -13,16 +13,17 @@ use andromeda_cw20_staking::CW20StakingContract;
 use andromeda_cw721::CW721Contract;
 use andromeda_distance::DistanceContract;
 use andromeda_economics::EconomicsContract;
+use andromeda_fixed_amount_splitter::FixedAmountSplitterContract;
 use andromeda_graph::GraphContract;
 use andromeda_ibc_registry::IBCRegistryContract;
 use andromeda_kernel::KernelContract;
 use andromeda_lockdrop::LockdropContract;
 use andromeda_marketplace::MarketplaceContract;
 use andromeda_merkle_airdrop::MerkleAirdropContract;
+use andromeda_point::PointContract;
 use andromeda_primitive::PrimitiveContract;
 use andromeda_rate_limiting_withdrawals::RateLimitingWithdrawalsContract;
 use andromeda_rates::RatesContract;
-use andromeda_set_amount_splitter::SetAmountSplitterContract;
 use andromeda_shunting::ShuntingContract;
 use andromeda_splitter::SplitterContract;
 use andromeda_std::deploy::ADOMetadata;
@@ -32,7 +33,6 @@ use andromeda_validator_staking::ValidatorStakingContract;
 use andromeda_vesting::VestingContract;
 use andromeda_vfs::VFSContract;
 use andromeda_weighted_distribution_splitter::WeightedDistributionSplitterContract;
-use andromeda_point::PointContract;
 
 use cw_orch::prelude::*;
 use cw_orch_daemon::{DaemonBase, Wallet};
@@ -71,7 +71,7 @@ pub fn all_contracts() -> Vec<DeployableContract> {
         deployable!(StringStorageContract),
         deployable!(ConditionalSplitterContract),
         deployable!(RateLimitingWithdrawalsContract),
-        deployable!(SetAmountSplitterContract),
+        deployable!(FixedAmountSplitterContract),
         deployable!(WeightedDistributionSplitterContract),
         deployable!(Cw20ExchangeContract),
         deployable!(LockdropContract),